Removed need to swimlaneDaD.js file, made some changes to the javascript to remove global variables
