+++ assets/helperFunctions.js
@@ -1,0 +1,5 @@
+export function setAttributes(element, attributes) {
+  for(var key in attributes) {
+    element.setAttribute(key, attributes[key])
+  }
+}
