+++ index.html
@@ -5,13 +5,10 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" type="text/css" href="index.css" />
   <script src="./index.js" defer type="module"></script>
-  <!-- <script src="./swimlaneListener.js" defer></script> -->
-  <script src="./swimlaneDaD.js" defer type="module"></script>
-  <!-- <script src="./todoListener.js" defer></script> -->
   <title>To-Done</title>
 </head>
 <body>
-  <h1>To-Done: Who are you?</h1>
+  <h1>To-Done: Drag me Crazy</h1>
   <div class="swimlaneContainer">
     <div class="swimlane newSwimlane">
       <button class="addSwimlaneButton">+ Add Swimlane</button>
